Support for accessing read-only repositories

+++ macrotests/regression-v1.sh
@@ -47,17 +47,21 @@ echo "--- Test future version detection"
 tar xzf $TESTHOME/regression-v1.tar.gz || exit 1
 echo "3" >$REPO/version.txt || exit 1
 REPO_PATH=$REPO $BOAR verify && { echo "Future version repo should fail"; exit 1; }
+(REPO_PATH=$REPO $BOAR verify 2>&1 | grep "Repo is from a future boar version.") || \
+    { echo "Operation didn't give expected error message"; exit 1; }
 rm -r regression-v1 || exit 1
 
 echo "--- Test write protected repo"
-# Boar does not yet handle write-protected repos. Make sure that such
-# repos are detected and handled gracefully.
+# Boar does not yet handle write-protected repos with pending
+# changes. Make sure that such repos are detected and handled
+# gracefully.
 tar xzf $TESTHOME/regression-v1.tar.gz || exit 1
-chmod -R a-w regression-v1 || exit 1
-REPO_PATH=$REPO $BOAR verify && { echo "Operation expected to fail due to write protect"; exit 1; }
-(REPO_PATH=$REPO $BOAR verify 2>&1 | grep "The repository seems to be read-only") || \
+test -e $REPO/queue/7 || { echo "Tested repo must contain an uncompleted commit"; exit 1; }
+chmod -R a-w $REPO || exit 1
+REPO_PATH=$REPO $BOAR verify && { echo "Operation expected to fail due to write protect and pending changes."; exit 1; }
+(REPO_PATH=$REPO $BOAR verify 2>&1 | grep "Repo is write protected with pending changes. Cannot continue.") || \
     { echo "Operation didn't give expected error message"; exit 1; }
-chmod -R u+w regression-v1 || exit 1
+chmod -R u+w $REPO || exit 1
 rm -r regression-v1 || exit 1
 
 echo "--- Test repo cloning"
